pyproject.toml
[project]
name = "LESC"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "vlad boiko", email = "acc.v.boiko@gmail.com"},
]
dependencies = [
    "fastapi>=0.115.3",
    "dependency-injector>=4.42.0",
    "pytest>=8.3.3",
    "alembic>=1.13.3",
    "pydantic-settings>=2.6.1",
    "uvicorn>=0.32.0",
    "pyjwt>=2.9.0",
    "bcrypt>=4.2.0",
    "tenacity>=9.0.0",
    "opentelemetry-distro>=0.48b0",
    "opentelemetry-exporter-otlp>=1.27.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.27.0",
    "opentelemetry-instrumentation-requests>=0.48b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.48b0",
    "psycopg2-binary>=2.9.10",
    "psycopg2>=2.9.10",
    "python-multipart>=0.0.17",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
